backend/requirements.txt
fastapi==0.115.0
pydantic-settings==2.5.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
anthropic==0.34.2

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Task queue
celery==5.4.0
redis==5.1.0

# File conversion
python-docx==1.1.2
pdfplumber==0.11.4
pdf2docx==0.5.8
weasyprint==62.3
Pillow==10.4.0
markdown==3.7
openpyxl==3.1.5
pandas==2.2.3
pypdf==5.1.0
python-pptx==1.0.2

# Storage
boto3==1.35.40

# Auth
authlib==1.3.2
itsdangerous==2.2.0

# Utils
python-dotenv==1.0.1
aiofiles==24.1.0
loguru==0.7.2
slowapi==0.1.9